Bug fixing.
The compiler crashed on 447.dealII and 483.xalancbmk because of
incomplete definition of structures required by certain pure virtual
functions. I added a check so if the definitions are incomplete,
the pure virtual function is not processed. Since the definitions
are incomplete, no member pointer can be defined, so we are safe.