Runtime support for virtual method address taken by entering the corresponding class's constructor for the first time.
Note: if a polymorphic class does not have constructor, then its
virtual methods' address are taken at the beginning of a program.
Check 473.astar regwayobj for more details.

Note': class construction without invoking constructors, for example,
populate a memory region, which has the same result as a polymorphic
class object, and then cast that region's pointer to a class object.
Such cases should be very rare and should be able to be detected
statically.